Industry666
This "Industry666" directory is an integral part of the project of the same name, which aims to explore the possibilities of Industry 6.6.6, merging technology and biology. Inside this directory, you will find the "industry666.py" file, which contains an initial code draft to illustrate the project's vision. The README.md file provides additional information on using the code and the next steps of the project.

Purpose and Objectives:

This directory serves as a starting point for developers interested in contributing to the vision of Industry 6.6.6.
The "industry666.py" file provides a basic implementation of the Industry 6.6.6 concept, which can be further developed and customized.

Detailed Contents:

"industry666.py" contains a basic class representing the Industry 6.6.6 concept, with methods for displaying the vision, pillars, and applications.

"README.md" provides instructions for running the code and additional project information.

Instructions for Use:

To run the code, follow the instructions in the README.md file.

Contributions and Collaborations:
We are open to contributions and collaborations from other developers interested in expanding this initial draft.

License:
This project is distributed under the MIT License.

For more information about the project and to contribute, visit https://www.rigeneproject.org/industry-6-6-6.